+++ Billing.API/Models/StockModel.cs
@@ -13,9 +13,9 @@ namespace Billing.API.Models
             public string Name;
         }
         public int Id { get; set; }
-        public int Input { get; set; }
-        public int Output { get; set; }
-        public int Inventory { get { return Input - Output; } }
+        public double Input { get; set; }
+        public double Output { get; set; }
+        public double Inventory { get { return Input - Output; } }
         public StockProduct Product { get; set; }
     }
 }
